_graphical_input: retry launch of keyboard if it fails
First attempt at this often fails for some reason - while we're
typing 'keyboard' the results come up as normal, but before we
hit enter, they all disappear and are replaced with "no results
found" (in Japanese). Dunno why. This will hopefully work around
that, if it works reliably on the second try.

Signed-off-by: Adam Williamson <awilliam@redhat.com>