Remove Plymouth from Server install_default_upload (#1933378)
This is a workaround for
https://bugzilla.redhat.com/show_bug.cgi?id=1933378 . Sometimes
when booting Server on a tap test without a working network
config (no DHCP server, static network not yet configured) we
hit a bug where the splash screen does not clear completely, and
this causes all the console needle matches to fail. To work
around this, we remove plymouth from the installed system after
running the install_default_upload test on Server; all affected
tests use the image uploaded by that test. We exclude aarch64
because there's a known problem with removing plymouth on that
arch (#1940163), plus the bug doesn't actually seem to happen on
aarch64 for some reason.

Signed-off-by: Adam Williamson <awilliam@redhat.com>