Add variant anaconda layout indicator needles for F35 stable
This seems to happen every cycle: when we get to testing an
RC, the layout indicators in anaconda move a bit (the text
moves relative to the little keyboard icon). I think the fact
that the prerelease text is no longer present is what causes
this. Anyway, we need to handle it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>