Revert "Fix the os_release.pm and fedora_release.pm."
This reverts commit 182bfdaa55a37549ebccfa471ffbb8fb684bde9d. The
problem was not in the test code here, it's just that the tests
were run with RAWREL set to 36 when it should be 37; this is
because fedfind's data source hadn't been updated for 36 branch
when the tests ran. It's tricky to time this exactly right -
ideally we'd probably update fedfind's data source one second
before the first Branched compose completes, but of course that's
difficult in practice.