Add use_cups_printer needle for F35, rearrange needles a bit
We should not use the same name for two different needles even
in two different directories as it can be confusing in some parts
of the UI which don't account for the directory name. Let's use
names differentiated by desktop. Also add a needle for F35 as
the one from the PR doesn't match (different relative placement
of icon and text).

Signed-off-by: Adam Williamson <awilliam@redhat.com>