printing: always use ls to find PDF, update needles
When the printing_builtin test ran on an F35 respin compose it
failed; it turns out the target filename was different for the
built-in print-to-PDF on GNOME on F35. So let's just always
use the 'ls' output to find the file, but pick the directory
to check based on whether we're using cups or not.

Also rename the needles to have unique names, and add one for
F35 GNOME.

Signed-off-by: Adam Williamson <awilliam@redhat.com>