Update a needle and fix launching keyboard settings on Japanese
To get to the keyboard/input method settings and add an input
method when doing a Japanese install test, we type 'keyboard',
but in current GNOME 42.beta that doesn't find the right pane.
Typing 'input' does work, though, so let's use that instead.
Also the GDM login needle needed updating.

Signed-off-by: Adam Williamson <awilliam@redhat.com>