Add needles for old flatpak gnome-text-editor
Silverblue has an older version of gnome-text-editor, using a
Flatpak base environment built off an older GNOME. Because of
this most of the needles for the current RPM-packaged version
don't match. For some reason the old needles we delete don't
match either - some difference in font rendering configuration
or something. So I had to create a bunch of new needles.

Signed-off-by: Adam Williamson <awilliam@redhat.com>