Update dnfdragora_runs for over-large window on KDE
The dnfdragora window is too big on current KDE so we can't see
the top of it (where we used to match). It's kinda better to
match on some active element of the app than just the window
title anyway (so we don't 'pass' if the window loads but is
empty, or something like that).

Signed-off-by: Adam Williamson <awilliam@redhat.com>