Try running iot_zezere_server on minimal UEFI image instead
OK, so we can't have parallel tests on different machines, so
we can't have iot_zezere_server on 64bit and iot_zezere_ignition
on uefi. We need both to be on uefi. So, let's try using the
minimal-uefi image that we build for upgrade tests instead. It's
built for -1 (which is the same as CURRREL) on x86_64 and aarch64
so it should be there when we need it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>