Update variant anaconda layout needles for F36 RCs
We get these variant needles with the keyboard icon and layout
name at slightly different relative vertical positions every
time we do a Final RC. I think the layout is affected by the
pre-release warning text no longer being present. Usually each
cycle something changes with the icon or with font rendering,
so the needles have to be updated, as in this case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>