desktop_login: drop obsolete check for 'login failed' message
Up to F33 we could check that an error message was shown when we
tried to log in with the wrong password on GNOME, but since F34
it's transient and disappears too quick to reliably catch, so
we don't check it any more. Now F33 is EOL, drop the conditional.

Signed-off-by: Adam Williamson <awilliam@redhat.com>