Make firefox nobookmarks needle not workaround any more
We've been told the bookmark toolbar no longer being shown on app
start is by design, so we can't call this a workaround needle
any more really. There's a preference we could use to turn it
back on so we can wait for the bookmark toolbar to show up to
be sure firefox is fully started, but in fact tests seem to be
passing OK with just this needle these days, so we may not need
to.

Signed-off-by: Adam Williamson <awilliam@redhat.com>