check_default_fonts: move reference files, use diff -u
Let's not have the reference files in one person's fedorapeople
space, in case that person leaves. Let's upload the text.txt
for checking (it's easier to be able to just read what's in it
than try and figure it out from the diff output), and let's use
diff -u because non-unified diff output is awful.

Signed-off-by: Adam Williamson <awilliam@redhat.com>