Drop NUMDISKS=2 for update server flavor
It causes a bit of an awkward problem for tests which use disk
images from another test, specifically the cockpit tests. We
put the update repo on this second disk and update /etc/fstab
but we aren't actually uploading the second disk image and using
it as the second disk on the child tests, so they get messed
up.

I'm having trouble coming up with an elegant solution so for now
let's kick the affected flavor (server) back to one disk. I'll
try and figure a more permanent fix tomorrow.

Signed-off-by: Adam Williamson <awilliam@redhat.com>