Protect against long load times in Cockpit
We seem to be hitting very long loads on the Services and Logs
pages of Cockpit in recent Rawhide testing especially. As I don't
have time to deeply debug this at the moment, let's just give it
longer (but make it a soft failure when it takes longer than
expected).

Signed-off-by: Adam Williamson <awilliam@redhat.com>